requirements.txt
pandas==1.5.3
Pillow==9.4.0
streamlit==1.23.1
scikit_learn==1.2.2
xgboost==1.7.1
